backend/requirements.txt
flask==3.0.2
flask-socketio==5.3.6
kafka-python==2.0.2
python-dotenv==1.0.1
fastavro==1.9.4

psycopg2-binary==2.9.9
SQLAlchemy==2.0.27